Moved callback helpers to a separate file
+ fixed a typo and minorly improved their usage in main
